restructured data and added equipment data
