+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "chartrack",
-  "version": "0.1.0",
+  "version": "0.1.1",
   "private": true,
   "dependencies": {
     "@material-ui/core": "^4.5.2",
